added whitespaceAround: '+'.
