--- src/main/java/core/basesyntax/Model/Cat.java
@@ -1,16 +1,0 @@
-package core.basesyntax.Model;
-
-public class Cat {
-    private int age;
-    private String name;
-    private String ownerName;
-
-    @Override
-    public String toString() {
-        return "Cat{" +
-                "age=" + age +
-                ", name='" + name + '\'' +
-                ", ownerName='" + ownerName + '\'' +
-                '}';
-    }
-}
